put FuelMeteringFuncTranslator defs into their own submodule
